productivity_app-master/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
python-dotenv==1.0.0
firebase-admin==6.2.0
Pyrebase4==4.7.1
gunicorn==21.2.0
Werkzeug==2.3.7 